backend/api/pyproject.toml
[project]
name = "sentinel-api"
version = "0.3.0"
description = "Sentinel FastAPI Lambda"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "mangum>=0.19.0",
  "pydantic>=2.7.0",
  "boto3>=1.34.0",
  "PyJWT[crypto]>=2.8.0",
]
